refactor: move functions out from utils.go (#1139)
